requirements.txt
yfinance
pandas
numpy
fastapi
uvicorn
psycopg2-binary
requests
beautifulsoup4
pydantic
